Refactor and remove utils
