+++ packages/i18n-core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@jupri-lab/i18n-core",
-  "version": "2.1.2",
+  "version": "2.1.3",
   "module": "dist/index.js",
   "types": "dist/index.d.ts",
   "type": "module",
