+++ packages/i18n-core/src/LanguageManager.ts
@@ -1,5 +1,4 @@
 import type EventManager from "./EventManager";
-import { appendQueryParams } from "./utils/appendQueryParams.util";
 
 export default class LanguageManager<TLanguage extends string> {
   private languages: TLanguage[];
@@ -37,8 +36,15 @@ export default class LanguageManager<TLanguage extends string> {
   changeLanguage(language: TLanguage, eventManager: EventManager<string>) {
     if (this.currentLanguage !== language) {
       this.currentLanguage = language;
-      if (this.enableQueryParams) appendQueryParams(this.queryParam, language);
+      if (this.enableQueryParams)
+        this.appendQueryParams(this.queryParam, language);
       eventManager.trigger("language_changed");
     }
+  }
+
+  private appendQueryParams(key: string, value: string) {
+    const newURL = new URL(window.location.href);
+    newURL.searchParams.set(key, value);
+    window.history.replaceState({}, "", newURL.href);
   }
 }
