--- packages/i18n-core/src/utils/escapeHTMLTags.util.ts
@@ -1,16 +1,0 @@
-export const escapeHTMLTags = (string: string): string => {
-  return string.replace(/[&<>"'`=/]/g, (match) => {
-    const escapeMap: Record<string, string> = {
-      // eslint-disable-next-line quotes
-      '"': "&quot;",
-      "&": "&amp;",
-      "'": "&#39;",
-      "/": "&#47;",
-      "<": "&lt;",
-      "=": "&#61;",
-      ">": "&gt;",
-      "`": "&#96;",
-    };
-    return escapeMap[match];
-  });
-};
